Feature: add webpack build to service worker
