+++ web/public/service-worker.js
@@ -1,70 +1,1 @@
-// declare const self: ServiceWorkerGlobalScope;
-
-/**
- * Array to store all the connected ports in.
- */
-const connectedPorts = [];
-
-// Create socket instance.
-const baseUrl = `ws://localhost:9003/api/v2/ranger/public`;
-const ws = new WebSocket(baseUrl);
-
-ws.onopen = () => {
-  const subscribe = {
-    event: 'subscribe',
-    streams: [
-      'global.tickers',
-      'dashbtc.trades'
-    ],
-  };
-
-  console.log('[Service worker] - Connection opened');
-
-  ws.send(JSON.stringify(subscribe));
-};
-
-ws.onmessage = (event) => {
-  console.log('[Service worker] - New message: ', event);
-  //const response = JSON.parse(event.data);
-  //setOrders(response.data);
-};
-
-ws.onclose = () => {
-  console.log('[Service worker] - Connection closed');
-
-  ws.close();
-};
-
-
-/*
-// Send data from socket to all open tabs.
-ws.addEventListener('message', ({ data }) => {
-  const payload = JSON.parse(data);
-  connectedPorts.forEach(port => port.postMessage(payload));
-});
-
-
-self.addEventListener('connect', ({ ports }: any) => {
-  const port = ports[0];
-
-  // Add this new port to the list of connected ports.
-  connectedPorts.push(port);
-
-  port.addEventListener('message', ({ data }) => {
-    const { action, value } = data;
-
-    // Send message to socket.
-    if (action === 'send') {
-      ws.send(JSON.stringify(value));
-
-    // Remove port from connected ports list.
-    } else if (action === 'unload') {
-      const index = connectedPorts.indexOf(port);
-      connectedPorts.splice(index, 1);
-    }
-  });
-
-  // Start the port broadcasting.
-  port.start();
-});
-*/
+!function(e){var n={};function o(r){if(n[r])return n[r].exports;var t=n[r]={i:r,l:!1,exports:{}};return e[r].call(t.exports,t,t.exports,o),t.l=!0,t.exports}o.m=e,o.c=n,o.d=function(e,n,r){o.o(e,n)||Object.defineProperty(e,n,{enumerable:!0,get:r})},o.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},o.t=function(e,n){if(1&n&&(e=o(e)),8&n)return e;if(4&n&&"object"==typeof e&&e&&e.__esModule)return e;var r=Object.create(null);if(o.r(r),Object.defineProperty(r,"default",{enumerable:!0,value:e}),2&n&&"string"!=typeof e)for(var t in e)o.d(r,t,function(n){return e[n]}.bind(null,t));return r},o.n=function(e){var n=e&&e.__esModule?function(){return e.default}:function(){return e};return o.d(n,"a",n),n},o.o=function(e,n){return Object.prototype.hasOwnProperty.call(e,n)},o.p="/",o(o.s=0)}([function(e,n,o){e.exports=o(1)},function(e,n){const o=new WebSocket("ws://localhost:9003/api/v2/ranger/public");o.onopen=()=>{console.log("[Service worker] - Connection opened"),o.send(JSON.stringify({event:"subscribe",streams:["global.tickers","dashbtc.trades"]}))},o.onmessage=e=>{console.log("[Service worker] - New message: ",e)},o.onclose=()=>{console.log("[Service worker] - Connection closed"),o.close()}}]);
